Extract SoundTagManager from Embed
