+++ include/ui/MenuBar.h
@@ -4,6 +4,7 @@
 #include <gtkmm.h>
 
 class MenuBar {
+  private:
   Glib::RefPtr<Gtk::Builder> m_refBuilder;
   Gtk::Box m_Box; 
 
@@ -12,6 +13,15 @@ class MenuBar {
   Glib::RefPtr<Gio::SimpleAction> m_refChoiceOther;
 
   Glib::RefPtr<Gio::SimpleAction> m_refToggle;
+
+  public:
+  //Signal handlers:
+  void on_menu_others();
+
+  void on_menu_choices(const Glib::ustring& parameter);
+  void on_menu_choices_other(int parameter);
+  void on_menu_toggle();
+
 };
 
 #endif
